% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_piece.r
\name{str_piece}
\alias{str_piece}
\title{Generate plaintext piecepack diagrams}
\usage{
str_piece(
  df,
  color = NULL,
  reorient = "none",
  annotate = FALSE,
  ...,
  annotation_scale = NULL,
  style = c("Unicode", "Game Bit Mono", "Game Bit Duo"),
  xbreaks = NULL,
  ybreaks = NULL
)
}
\arguments{
\item{df}{Data frame containing piece info.}

\item{color}{How should the text be colorized.
If \code{FALSE} won't colorize output at all.
If \code{"html"} will colorize output for html.
Otherwise will colorize output for the terminal using ANSI CSI SGR control sequences.}

\item{reorient}{Determines whether and how we should reorient (the angle) of pieces or symbols:\enumerate{
\item{The default "none" (or \code{FALSE}) means don't reorient any pieces/symbols.}
\item{"all" (or \code{TRUE}) means setting the angle to zero for all pieces
(reorienting them all \dQuote{up}).}
\item{"symbols" means just re-orient suit/rank symbols but not the orientation of the piece itself.
In particular, in contrast with "all" this preserves the location
of the upper-left "corner" of piecepack tile faces.}}}

\item{annotate}{If \code{TRUE} or \code{"algebraic"} annotate the plot
with \dQuote{algrebraic} coordinates,
if \code{FALSE} or \code{"none"} don't annotate,
if \code{"cartesian"} annotate the plot with \dQuote{cartesian} coordinates.}

\item{...}{Mainly ignored except for a couple of undocumented features.}

\item{annotation_scale}{Multiplicative factor that scales (stretches) any annotation coordinates.
By default uses \code{attr(df, "scale_factor") \%||\% 1}.}

\item{style}{If "Unicode" (default) only use glyphs in Unicode proper.
If "Game Bit Duo" use glyphs in Private Use Area of "Game Bit Duo" font.
If "Game Bit Mono" use glyphs in Private Use Area of "Game Bit Mono" font.}

\item{xbreaks, ybreaks}{Subset (of integers) to provide axis labels for if \code{annotate} is \code{TRUE}.
If \code{NULL} infer a reasonable choice.}
}
\value{
Character vector for text diagram.
}
\description{
\code{str_piece()} generates plaintext piecepack diagrams.
}
\seealso{
\code{\link[=cat_piece]{cat_piece()}} for printing to the terminal.
See \url{https://github.com/trevorld/game-bit-font} for more information about the \dQuote{Game Bit} family of fonts.
}
